pyproject.toml
[project]
name = "simple-streamlit-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "compass-api-sdk>=0.3.5",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "streamlit-autorefresh>=1.0.1",
    "web3>=7.11.1",
]


[dependency-groups]
dev = [
    "httpx >=0.28.1,<1.0.0",
    "pytest-timeout >=2.3.1,<3.0.0",
    "black >=24.10.0,<25.0.0",
    "pre-commit >=4.0.1,<5.0.0",
    "isort >=5.13.2,<6.0.0",
    "docformatter >=1.7.5,<2.0.0",
    "flake8 >=7.1.1,<8.0.0",
    "pydocstyle >=6.3.0,<7.0.0",
    "mypy >=1.13.0,<2.0.0",
    "ruff >=0.9.7,<1.0.0"
]

[tool.setuptools]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

#[tool.mypy]
#plugins = ["pydantic.mypy"]

[tool.pyright]
typeCheckingMode = "strict"
#reportMissingParameterType="none"
reportUnknownVariableType="none"
reportUnknownMemberType="none"
reportUnknownArgumentType="none"
reportUnknownParameterType="none"
reportUnnecessaryComparison="none"
reportUnnecessaryIsInstance="none"
reportMissingTypeArgument="none"
reportPrivateUsage="none"
reportArgumentType="none"
reportUntypedFunctionDecorator="none"
reportDeprecated="none"
reportMissingTypeStubs="none"
reportUnusedVariable="none"
reportIncompatibleMethodOverride="none"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
